refactor: decompose Skill Atlas UI and add ESLint
- Split the 963-line App.tsx into focused modules (panels, inspector,
  route preview, graph nodes, primitives, helpers); App is now an orchestrator
- Memoize routeOrder and the DomainInspector edge scan; split the skill-detail
  fetch effect from selection persistence so an atlas refresh does not refetch
- Make the dev-proxy port configurable via SKILLROUTE_UI_PORT
- Add a flat ESLint config (react-hooks rules) and a lint script
