+++ web/src/ui.ts
@@ -1,0 +1,2 @@
+export type InspectorTab = "overview" | "skills" | "excerpts" | "stats";
+export type NavSection = "map" | "atlas" | "routes" | "catalog";
